Add files via upload
The file "Preliminary _thesis_Di_Giacomo" is a first draft of the Thesis Project. Currently working on the Introduction part, and on the graphics of the results. 

Abstract:

As humanity embarks on a quest for sustainable energy solutions, nuclear fusion is gaining momentum, emerging as a beacon of hope and offering transformative solutions to pressing environmental challenges and concerns. This work provides an overview of the fundamentals of nuclear fusion, its historical development, and the current state of research. Furthermore, it focuses on one of the critical aspects of fusion energy development: material selection. The success of fusion as a viable energy source hinges on the development of materials capable of withstanding the extreme conditions within a fusion reactor, including high temperatures, intense radiation, corrosive environments and mechanical loads. Particular attention has been devoted to one of the most critical and expensive components of the reactor, namely the superconducting magnets. More precisely this work delves on the toroidal field shield optimization. Variation of the Tritium breeding ratio (TBR) when different materials are used as shielding of the Toroidal Field Coil (TFC) in an ARC like fusion reactor has been confirmed resulting in a reduction of the TBR when neutron flux increases on the high temperature superconductors (HTS). TiH2 , for some reactor design configurations, and tungsten carbide exhibit superior shielding properties compared to the actual baseline material ZrH2 . Furthermore FLiPb has been investigated as coolant and blanket materials leading to a significant reduction of the capital cost of the reactor compared to FLiBe. Finally activation analysis on different vacuum vessel materials (V4Cr4Ti, F82H, and SiC) have been performed, including shutdown dose rate (SDR) based on the Rigorous 2-Step (R2S) approach for the chromium vanadium alloy. To this purpose have been evaluated integral fluxes, neutron and PKA spectra, dpa, specific activities, dose rates and TBR for different ARC design reactor configurations.